pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "cogcc-pipeline"
version = "0.1.0"
description = "COGCC/ECMC oil and gas production data pipeline"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.28.0",
    "beautifulsoup4>=4.11.0",
    "dask[complete]>=2022.12.0",
    "distributed>=2022.12.0",
    "bokeh>=2.4.0",
    "pyyaml>=6.0",
    "tenacity>=8.1.0",
    "tqdm>=4.64.0",
    "pandas>=1.5.0",
    "pyarrow>=10.0.0",
    "numpy>=1.23.0",
    "lxml>=4.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.2.0",
    "pytest-mock>=3.10.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
cogcc-pipeline = "cogcc_pipeline.pipeline:main"

[tool.hatch.build.targets.wheel]
packages = ["cogcc_pipeline"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.mypy]
python_version = "3.10"
ignore_missing_imports = true
warn_unused_ignores = false
